pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["aimee aimeemiranribe@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pymongo = "^4.11"
uvicorn = "^0.34.0"
python-decouple = "^3.8"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["dotenv"], version = "^2.10.6"}
python-multipart = "^0.0.20"

[tool.setuptools]
packages = [".", "src"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
